+++ openpifpaf/plugins/coco/cocokp.py
@@ -220,10 +220,13 @@ class CocoKp(openpifpaf.datasets.DataModule, openpifpaf.Configurable):
             rescale_t,
             openpifpaf.transforms.RandomApply(
                 openpifpaf.transforms.Blur(), self.blur),
+            openpifpaf.transforms.RandomChoice(
+                [openpifpaf.transforms.RotateBy90(),
+                 openpifpaf.transforms.RotateUniform(30.0)],
+                [self.orientation_invariant, 0.4],
+            ),
             openpifpaf.transforms.Crop(self.square_edge, use_area_of_interest=True),
             openpifpaf.transforms.CenterPad(self.square_edge),
-            openpifpaf.transforms.RandomApply(
-                openpifpaf.transforms.RotateBy90(), self.orientation_invariant),
             openpifpaf.transforms.TRAIN_TRANSFORM,
             openpifpaf.transforms.Encoders(encoders),
         ])