+++ openpifpaf/visualizer/__init__.py
@@ -1,10 +1,19 @@
 """Higher level drawing functions."""
 
 from .base import Base
+from .caf import Caf
 from .cif import Cif
 from .cifdet import CifDet
 from .cifhr import CifHr
 from .cli import cli, configure
-from .caf import Caf
 from .occupancy import Occupancy
 from .seeds import Seeds
+
+__all__ = [
+    'Caf',
+    'Cif',
+    'CifDet',
+    'CifHr',
+    'Occupancy',
+    'Seeds',
+]